requirements.txt
#numpy==1.19.4

pyserial~=3.5
numpy~=1.19.4
opencv-python~=4.4.0.46